Move wall code to it's own file
