+++ src/map/helpers/zone_generation/mod.rs
@@ -2,6 +2,6 @@ mod create_tile_layout;
 mod hub_generator;
 mod instance_generator;
 mod map_data;
-
+mod walls;
 pub use hub_generator::generate_hub_layout;
 pub use instance_generator::generate_instance_layout;
